+++ app/lib/gen/strings_pl.g.dart
@@ -142,7 +142,7 @@ class _TranslationsGeneralPl extends TranslationsGeneralEn {
   @override
   String get quickSave => 'Szybki zapis';
   @override
-  String get quickSaveFromFavorites => 'Szybki zapis dla "Ulubionych"';
+  String get quickSaveFromFavorites => 'Szybki zapis dla „Ulubionych”';
   @override
   String get renamed => 'Przemianowany';
   @override
@@ -180,6 +180,8 @@ class _TranslationsReceiveTabPl extends TranslationsReceiveTabEn {
   String get title => 'Odbierz';
   @override
   late final _TranslationsReceiveTabInfoBoxPl infoBox = _TranslationsReceiveTabInfoBoxPl._(_root);
+  @override
+  late final _TranslationsReceiveTabQuickSavePl quickSave = _TranslationsReceiveTabQuickSavePl._(_root);
 }
 
 // Path: sendTab
@@ -376,7 +378,7 @@ class _TranslationsSendPagePl extends TranslationsSendPageEn {
 
   // Translations
   @override
-  String get waiting => 'Czekanie na odpowiedź...';
+  String get waiting => 'Czekanie na odpowiedź…';
   @override
   String get rejected => 'Odbiorca odrzucił żądanie.';
   @override
@@ -416,9 +418,9 @@ class _TranslationsWebSharePagePl extends TranslationsWebSharePageEn {
   @override
   String get title => 'Udostępnij przez link';
   @override
-  String get loading => 'Uruchamianie serwera...';
+  String get loading => 'Uruchamianie serwera…';
   @override
-  String get stopping => 'Zatrzymywanie serwera...';
+  String get stopping => 'Zatrzymywanie serwera…';
   @override
   String get error => 'Wystąpił błąd podczas uruchamiania serwera.';
   @override
@@ -438,7 +440,7 @@ class _TranslationsWebSharePagePl extends TranslationsWebSharePageEn {
   @override
   String get requirePin => 'Wymagaj kodu PIN';
   @override
-  String pinHint({required Object pin}) => 'Kod PIN to "${pin}"';
+  String pinHint({required Object pin}) => 'Kod PIN to „${pin}”';
   @override
   String get encryptionHint => 'LocalSend używa samopodpisanego certyfikatu. Musisz go zaakceptować w przeglądarce.';
   @override
@@ -723,6 +725,23 @@ class _TranslationsReceiveTabInfoBoxPl extends TranslationsReceiveTabInfoBoxEn {
   String get alias => 'Alias:';
 }
 
+// Path: receiveTab.quickSave
+class _TranslationsReceiveTabQuickSavePl extends TranslationsReceiveTabQuickSaveEn {
+  _TranslationsReceiveTabQuickSavePl._(TranslationsPl root)
+      : this._root = root,
+        super.internal(root);
+
+  final TranslationsPl _root; // ignore: unused_field
+
+  // Translations
+  @override
+  String get off => _root.general.off;
+  @override
+  String get favorites => 'Ulubione';
+  @override
+  String get on => _root.general.on;
+}
+
 // Path: sendTab.selection
 class _TranslationsSendTabSelectionPl extends TranslationsSendTabSelectionEn {
   _TranslationsSendTabSelectionPl._(TranslationsPl root)
@@ -972,7 +991,7 @@ class _TranslationsTroubleshootPageNoConnectionPl extends TranslationsTroublesho
   String get symptom => 'Oba urządzenia nie mogą się nawzajem wykryć ani dzielić się plikami.';
   @override
   String get solution =>
-      'Problem istnieje po obu stronach? Upewnij się, że oba urządzenia są w tej samej sieci Wi-Fi i mają takie same ustawienia (port, adres multicast, szyfrowanie). Wi-Fi może nie zezwalać na komunikację między uczestnikami. W takim przypadku opcję tę należy włączyć w routerze.';
+      'Problem istnieje po obu stronach? Upewnij się, że oba urządzenia są w tej samej sieci Wi-Fi i mają takie same ustawienia (port, adres multicast, szyfrowanie). Wi-Fi może nie zezwalać na komunikację między uczestnikami z powodu izolacji punktu dostępu (AP). W takim przypadku tę opcję należy wyłączyć w routerze.';
 }
 
 // Path: receiveHistoryPage.entryActions
@@ -1089,7 +1108,7 @@ class _TranslationsDialogsCannotOpenFilePl extends TranslationsDialogsCannotOpen
   @override
   String get title => 'Nie można otworzyć pliku';
   @override
-  String content({required Object file}) => 'Nie można otworzyć "${file}". Czy ten plik został przeniesiony, zmieniono mu nazwę lub usunięto?';
+  String content({required Object file}) => 'Nie można otworzyć „${file}”. Czy ten plik został przeniesiony, zmieniono mu nazwę lub usunięto?';
 }
 
 // Path: dialogs.encryptionDisabledNotice
@@ -1149,7 +1168,7 @@ class _TranslationsDialogsFavoriteDeleteDialogPl extends TranslationsDialogsFavo
   @override
   String get title => 'Usuń z ulubionych';
   @override
-  String content({required Object name}) => 'Czy na pewno chcesz usunąć z ulubionych "${name}"?';
+  String content({required Object name}) => 'Czy na pewno chcesz usunąć z ulubionych „${name}”?';
 }
 
 // Path: dialogs.favoriteEditDialog
@@ -1368,6 +1387,12 @@ class _TranslationsDialogsQuickSaveFromFavoritesNoticePl extends TranslationsDia
   // Translations
   @override
   String get title => _root.general.quickSaveFromFavorites;
+  @override
+  List<String> get content => [
+        'Żądania plików są teraz akceptowane automatycznie z urządzeń na liście ulubionych.',
+        'Uwaga! Obecnie nie jest to całkowicie bezpieczne, ponieważ haker, który ma odcisk palca dowolnego urządzenia z listy ulubionych, może wysyłać Ci pliki bez ograniczeń.',
+        'Jednak ta opcja jest nadal bezpieczniejsza niż umożliwienie wszystkim użytkownikom w sieci lokalnej wysyłanie plików bez ograniczeń.',
+      ];
 }
 
 // Path: dialogs.pin