+++ app/lib/gen/strings_zh_CN.g.dart
@@ -140,9 +140,9 @@ class _TranslationsGeneralZhCn extends TranslationsGeneralEn {
   @override
   String get queue => '队列';
   @override
-  String get quickSave => '快速保存';
+  String get quickSave => '自动保存';
   @override
-  String get quickSaveFromFavorites => '快速保存到"收藏"';
+  String get quickSaveFromFavorites => '自动保存来自“收藏夹”设备的文件';
   @override
   String get renamed => '已重命名';
   @override
@@ -180,6 +180,8 @@ class _TranslationsReceiveTabZhCn extends TranslationsReceiveTabEn {
   String get title => '接收';
   @override
   late final _TranslationsReceiveTabInfoBoxZhCn infoBox = _TranslationsReceiveTabInfoBoxZhCn._(_root);
+  @override
+  late final _TranslationsReceiveTabQuickSaveZhCn quickSave = _TranslationsReceiveTabQuickSaveZhCn._(_root);
 }
 
 // Path: sendTab
@@ -466,7 +468,7 @@ class _TranslationsAboutPageZhCn extends TranslationsAboutPageEn {
   @override
   String get contributors => '贡献者';
   @override
-  String get packagers => '包';
+  String get packagers => '打包者';
   @override
   String get translators => '翻译者';
 }
@@ -600,6 +602,8 @@ class _TranslationsDialogsZhCn extends TranslationsDialogsEn {
   @override
   late final _TranslationsDialogsAddFileZhCn addFile = _TranslationsDialogsAddFileZhCn._(_root);
   @override
+  late final _TranslationsDialogsOpenFileZhCn openFile = _TranslationsDialogsOpenFileZhCn._(_root);
+  @override
   late final _TranslationsDialogsAddressInputZhCn addressInput = _TranslationsDialogsAddressInputZhCn._(_root);
   @override
   late final _TranslationsDialogsCancelSessionZhCn cancelSession = _TranslationsDialogsCancelSessionZhCn._(_root);
@@ -677,7 +681,7 @@ class _TranslationsTrayZhCn extends TranslationsTrayEn {
   @override
   String get close => '退出 LocalSend';
   @override
-  String get closeWindows => '退出';
+  String get closeWindows => '离开';
 }
 
 // Path: web
@@ -791,6 +795,23 @@ class _TranslationsReceiveTabInfoBoxZhCn extends TranslationsReceiveTabInfoBoxEn
   String get alias => '别名：';
 }
 
+// Path: receiveTab.quickSave
+class _TranslationsReceiveTabQuickSaveZhCn extends TranslationsReceiveTabQuickSaveEn {
+  _TranslationsReceiveTabQuickSaveZhCn._(TranslationsZhCn root)
+      : this._root = root,
+        super.internal(root);
+
+  final TranslationsZhCn _root; // ignore: unused_field
+
+  // Translations
+  @override
+  String get off => _root.general.off;
+  @override
+  String get favorites => '收藏夹';
+  @override
+  String get on => _root.general.on;
+}
+
 // Path: sendTab.selection
 class _TranslationsSendTabSelectionZhCn extends TranslationsSendTabSelectionEn {
   _TranslationsSendTabSelectionZhCn._(TranslationsZhCn root)
@@ -872,11 +893,11 @@ class _TranslationsSettingsTabGeneralZhCn extends TranslationsSettingsTabGeneral
   @override
   late final _TranslationsSettingsTabGeneralLanguageOptionsZhCn languageOptions = _TranslationsSettingsTabGeneralLanguageOptionsZhCn._(_root);
   @override
-  String get saveWindowPlacement => '关闭时保存窗口位置';
+  String get saveWindowPlacement => '退出时保存窗口位置';
   @override
-  String get saveWindowPlacementWindows => '退出时保存窗口位置';
+  String get saveWindowPlacementWindows => '离开时保存窗口位置';
   @override
-  String get minimizeToTray => '关闭时：最小化到系统托盘';
+  String get minimizeToTray => '关闭时最小化到系统托盘';
   @override
   String get launchAtStartup => '登录系统后自动启动程序';
   @override
@@ -1020,7 +1041,7 @@ class _TranslationsTroubleshootPageNoDiscoveryZhCn extends TranslationsTroublesh
   @override
   String get symptom => '此设备未能发现其他设备。';
   @override
-  String get solution => '确保所有设备都在同一个Wi-Fi网络上，并共享相同的配置（端口、多播地址、加密）。您可以尝试手动键入目标设备的IP地址。如果这有效，请考虑将此设备添加到收藏夹中，以便将来可以自动发现。';
+  String get solution => '确保所有设备都处于同一个 Wi‑Fi 网络上，且共享相同的网络配置（端口、多播地址、加密选项）。您可以尝试手动键入目标设备的 IP 地址。如果这有效，请考虑将此设备添加到收藏夹中，以便将来可以自动发现。';
 }
 
 // Path: troubleshootPage.noConnection
@@ -1035,7 +1056,7 @@ class _TranslationsTroubleshootPageNoConnectionZhCn extends TranslationsTroubles
   @override
   String get symptom => '双方设备均无法发现对方或者分享文件。';
   @override
-  String get solution => '当问题发生在双方设备上时，请先确认双方设备处于同一 Wi‑Fi 网络内，且有相同的网络（端口、多播地址、加密选项）配置。若因 Wi‑Fi 不允许参与者间通信，那么请在路由器中关闭这个（如：AP 隔离）选项。';
+  String get solution => '当问题发生在双方设备上时，请先确认双方设备处于同一个 Wi‑Fi 网络上，且共享相同的网络配置（端口、多播地址、加密选项）。若因 Wi‑Fi 不允许参与者间通信，那么请在路由器中关闭“接入点 (AP) 隔离”选项。';
 }
 
 // Path: receiveHistoryPage.entryActions
@@ -1089,6 +1110,21 @@ class _TranslationsDialogsAddFileZhCn extends TranslationsDialogsAddFileEn {
   String get title => '加入文件';
   @override
   String get content => '你想加入什么文件？';
+}
+
+// Path: dialogs.openFile
+class _TranslationsDialogsOpenFileZhCn extends TranslationsDialogsOpenFileEn {
+  _TranslationsDialogsOpenFileZhCn._(TranslationsZhCn root)
+      : this._root = root,
+        super.internal(root);
+
+  final TranslationsZhCn _root; // ignore: unused_field
+
+  // Translations
+  @override
+  String get title => '打开文件';
+  @override
+  String get content => '您是否要打开接收的文件？';
 }
 
 // Path: dialogs.addressInput
@@ -1415,6 +1451,12 @@ class _TranslationsDialogsQuickSaveFromFavoritesNoticeZhCn extends TranslationsD
   // Translations
   @override
   String get title => _root.general.quickSaveFromFavorites;
+  @override
+  List<String> get content => [
+        '当前会自动接受收藏夹中设备的文件请求。',
+        '警告：这目前并非绝对安全，知道您收藏夹列表中设备指纹的黑客仍可以向您发送文件。',
+        '但是，此选项比“允许任何设备”更安全。',
+      ];
 }
 
 // Path: dialogs.pin