+++ app/lib/pages/tabs/settings_tab.dart
@@ -15,6 +15,7 @@ import 'package:localsend_app/provider/settings_provider.dart';
 import 'package:localsend_app/provider/version_provider.dart';
 import 'package:localsend_app/util/alias_generator.dart';
 import 'package:localsend_app/util/device_type_ext.dart';
+import 'package:localsend_app/util/native/macos_channel.dart';
 import 'package:localsend_app/util/native/pick_directory_path.dart';
 import 'package:localsend_app/util/native/platform_check.dart';
 import 'package:localsend_app/widget/custom_dropdown_button.dart';
@@ -202,11 +203,17 @@ class SettingsTab extends StatelessWidget {
                       onPressed: () async {
                         if (vm.settings.destination != null) {
                           await ref.notifier(settingsProvider).setDestination(null);
+                          if (defaultTargetPlatform == TargetPlatform.macOS) {
+                            await removeExistingDestinationAccess();
+                          }
                           return;
                         }
 
                         final directory = await pickDirectoryPath();
                         if (directory != null) {
+                          if (defaultTargetPlatform == TargetPlatform.macOS) {
+                            await persistDestinationFolderAccess(directory);
+                          }
                           await ref.notifier(settingsProvider).setDestination(directory);
                         }
                       },