+++ app/lib/provider/network/send_provider.dart
@@ -1,8 +1,9 @@
 import 'dart:async';
+import 'dart:collection';
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:common/api_route_builder.dart';
+import 'package:common/isolate.dart';
 import 'package:common/model/device.dart';
 import 'package:common/model/dto/file_dto.dart';
 import 'package:common/model/dto/info_register_dto.dart';
@@ -13,7 +14,6 @@ import 'package:common/model/file_status.dart';
 import 'package:common/model/file_type.dart';
 import 'package:common/model/session_status.dart';
 import 'package:common/util/sleep.dart';
-import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:localsend_app/model/cross_file.dart';
 import 'package:localsend_app/model/send_mode.dart';
@@ -23,14 +23,14 @@ import 'package:localsend_app/pages/home_page.dart';
 import 'package:localsend_app/pages/progress_page.dart';
 import 'package:localsend_app/pages/send_page.dart';
 import 'package:localsend_app/provider/device_info_provider.dart';
-import 'package:localsend_app/provider/dio_provider.dart';
+import 'package:localsend_app/provider/http_provider.dart';
 import 'package:localsend_app/provider/progress_provider.dart';
 import 'package:localsend_app/provider/selection/selected_sending_files_provider.dart';
 import 'package:localsend_app/provider/settings_provider.dart';
-import 'package:localsend_app/util/stream.dart';
 import 'package:localsend_app/widget/dialogs/pin_dialog.dart';
 import 'package:logging/logging.dart';
 import 'package:refena_flutter/refena_flutter.dart';
+import 'package:rhttp/rhttp.dart';
 import 'package:routerino/routerino.dart';
 import 'package:uri_content/uri_content.dart';
 import 'package:uuid/uuid.dart';
@@ -62,7 +62,7 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
     required List<CrossFile> files,
     required bool background,
   }) async {
-    final requestDio = ref.read(dioProvider).longLiving;
+    final client = ref.read(httpProvider).longLiving;
     final cancelToken = CancelToken();
     final sessionId = _uuid.v4();
 
@@ -106,7 +106,7 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
       }))),
       startTime: null,
       endTime: null,
-      cancelToken: cancelToken,
+      sendingTasks: [],
       errorMessage: null,
     );
 
@@ -140,22 +140,23 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
       );
     }
 
-    Response? response;
+    HttpTextResponse? response;
     bool invalidPin;
     bool pinFirstAttempt = true;
     String? pin;
     do {
       invalidPin = false;
       try {
-        response = await requestDio.post(
-          ApiRoute.prepareUpload.target(target, query: {
+        response = await client.post(
+          ApiRoute.prepareUpload.target(target),
+          query: {
             if (pin != null) 'pin': pin,
-          }),
-          data: jsonEncode(requestDto), // jsonEncode for better logging
+          },
+          body: HttpBody.json(requestDto.toJson()),
           cancelToken: cancelToken,
         );
-      } on DioException catch (e) {
-        switch (e.response?.statusCode) {
+      } on RhttpStatusCodeException catch (e) {
+        switch (e.statusCode) {
           case 401:
             invalidPin = true;
 
@@ -234,7 +235,7 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
 
     final Map<String, String> fileMap;
     if (target.version == '1.0') {
-      fileMap = (response.data as Map).cast<String, String>();
+      fileMap = (response.bodyToJson as Map).cast<String, String>();
     } else {
       if (response.statusCode == 204) {
         // Nothing selected
@@ -242,7 +243,7 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
         fileMap = {};
       } else {
         try {
-          final responseDto = PrepareUploadResponseDto.fromJson(response.data);
+          final responseDto = PrepareUploadResponseDto.fromJson(response.bodyToJson);
           fileMap = responseDto.files;
           state = state.updateSession(
             sessionId: sessionId,
@@ -310,35 +311,55 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
       ),
     );
 
-    await _sendLoop(sessionId, target, sendingFiles);
+    await _sendLoop(ref, sessionId, target, sendingFiles);
   }
 
-  Future<void> _sendLoop(String sessionId, Device target, Map<String, SendingFile> files) async {
+  Future<void> _sendLoop(Ref ref, String sessionId, Device target, Map<String, SendingFile> files) async {
     state = state.updateSession(
       sessionId: sessionId,
       state: (s) => s?.copyWith(startTime: DateTime.now().millisecondsSinceEpoch),
     );
 
-    for (final file in files.values) {
-      final result = await sendFile(
-        sessionId: sessionId,
-        file: file,
-        isRetry: false,
-      );
-      if (!result) {
-        break;
+    final queue = Queue<SendingFile>()..addAll(files.values);
+    final concurrency = ref.read(parentIsolateProvider).uploadIsolateCount;
+    _logger.info('Sending files using $concurrency concurrent isolates');
+
+    final futures = List.generate(concurrency, (index) async {
+      while (true) {
+        final file = switch (queue.isEmpty) {
+          true => null,
+          false => queue.removeFirst(),
+        };
+
+        if (file == null) {
+          break;
+        }
+
+        await sendFile(
+          sessionId: sessionId,
+          isolateIndex: index,
+          file: file,
+          isRetry: false,
+        );
       }
-    }
+    });
+
+    await Future.wait(futures);
 
     _finish(sessionId: sessionId);
   }
 
   void _finish({required String sessionId}) {
-    if (state[sessionId] != null && state[sessionId]!.status != SessionStatus.sending) {
+    final sessionState = state[sessionId];
+    if (sessionState == null) {
+      return;
+    }
+
+    if (state[sessionId]!.status != SessionStatus.sending) {
       _logger.info('Transfer was canceled.');
     } else {
-      final hasError = state[sessionId]!.files.values.any((file) => file.status == FileStatus.failed);
-      if (!hasError && state[sessionId]!.background == true) {
+      final hasError = sessionState.files.values.any((file) => file.status == FileStatus.failed);
+      if (!hasError && sessionState.background == true) {
         // close session because everything is fine and it is in background
         closeSession(sessionId);
         _logger.info('Transfer finished and session removed.');
@@ -367,6 +388,7 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
   /// Returns true, if the next file should be sent.
   Future<bool> sendFile({
     required String sessionId,
+    required int isolateIndex,
     required SendingFile file,
     required bool isRetry,
   }) async {
@@ -382,7 +404,6 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
     }
 
     final remoteSessionId = state[sessionId]!.remoteSessionId;
-    final dio = ref.read(dioProvider).longLiving;
     final target = state[sessionId]!.target;
 
     if (isRetry) {
@@ -409,47 +430,38 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
       state: (s) => s?.withFileStatus(file.file.id, FileStatus.sending, null),
     );
 
-    final Stream<List<int>>? fileStream = file.path != null
-        ? file.path!.startsWith('content://')
-            ? uriContent.getContentStream(Uri.parse(file.path!))
-            : File(file.path!).openRead()
-        : null;
-
-    final (streamController, subscription) = fileStream?.digested() ?? (null, null);
+    final taskResult = ref.redux(parentIsolateProvider).dispatchTakeResult(IsolateHttpUploadAction(
+          isolateIndex: isolateIndex,
+          remoteSessionId: remoteSessionId,
+          remoteFileToken: token,
+          fileId: file.file.id,
+          filePath: file.path,
+          fileBytes: file.bytes,
+          mime: file.file.lookupMime(),
+          fileSize: file.file.size,
+          device: target,
+        ));
 
     String? fileError;
     try {
-      final cancelToken = CancelToken();
       state = state.updateSession(
         sessionId: sessionId,
-        state: (s) => s?.copyWith(cancelToken: cancelToken),
+        state: (s) => s?.copyWith(sendingTasks: [
+          ...?s.sendingTasks,
+          SendingTask(
+            isolateIndex: isolateIndex,
+            taskId: taskResult.taskId,
+          ),
+        ]),
       );
-      final stopwatch = Stopwatch()..start();
-      await dio.post(
-        ApiRoute.upload.target(target, query: {
-          if (remoteSessionId != null) 'sessionId': remoteSessionId,
-          'fileId': file.file.id,
-          'token': token,
-        }),
-        options: Options(
-          headers: {
-            'Content-Length': file.file.size,
-            'Content-Type': file.file.lookupMime(),
-          },
-        ),
-        data: streamController?.stream ?? file.bytes!,
-        onSendProgress: (curr, total) {
-          if (stopwatch.elapsedMilliseconds >= 100) {
-            stopwatch.reset();
-            ref.notifier(progressProvider).setProgress(
-                  sessionId: sessionId,
-                  fileId: file.file.id,
-                  progress: curr / total,
-                );
-          }
-        },
-        cancelToken: cancelToken,
-      );
+
+      await for (final progress in taskResult.progress) {
+        ref.notifier(progressProvider).setProgress(
+              sessionId: sessionId,
+              fileId: file.file.id,
+              progress: progress,
+            );
+      }
 
       // set progress to 100% when successfully finished
       ref.notifier(progressProvider).setProgress(
@@ -461,13 +473,11 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
       fileError = e.humanErrorMessage;
       _logger.warning('Error while sending file ${file.file.fileName}', e, st);
     } finally {
-      // Close the stream if it is still open
-      // ignore: unawaited_futures
-      streamController?.close();
-
-      // Cancel the subscription if it is still open
-      // ignore: unawaited_futures
-      subscription?.cancel();
+      state = state.updateSession(
+        sessionId: sessionId,
+        state: (s) => s?.copyWith(
+            sendingTasks: s.sendingTasks?.where((task) => !(task.isolateIndex == isolateIndex && task.taskId == taskResult.taskId)).toList()),
+      );
     }
 
     state = state.updateSession(
@@ -493,12 +503,13 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
       return;
     }
     final remoteSessionId = sessionState.remoteSessionId;
-    sessionState.cancelToken?.cancel(); // cancel current request
+
+    _cancelRunningRequests(sessionState);
 
     // notify the receiver
     try {
       ref
-          .read(dioProvider)
+          .read(httpProvider)
           .discovery
           // ignore: discarded_futures
           .post(ApiRoute.cancel.target(sessionState.target, query: remoteSessionId != null ? {'sessionId': remoteSessionId} : null));
@@ -515,7 +526,7 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
     if (sessionState == null) {
       return;
     }
-    sessionState.cancelToken?.cancel(); // cancel current request
+    _cancelRunningRequests(sessionState);
 
     state = state.updateSession(
       sessionId: sessionId,
@@ -524,6 +535,15 @@ class SendNotifier extends Notifier<Map<String, SendSessionState>> {
         endTime: DateTime.now().millisecondsSinceEpoch,
       ),
     );
+  }
+
+  void _cancelRunningRequests(SendSessionState state) {
+    for (final task in state.sendingTasks ?? <SendingTask>[]) {
+      ref.redux(parentIsolateProvider).dispatch(IsolateHttpUploadCancelAction(
+            isolateIndex: task.isolateIndex,
+            taskId: task.taskId,
+          ));
+    }
   }
 
   /// Closes the session
@@ -588,17 +608,27 @@ extension on SendSessionState {
 extension on Object {
   String get humanErrorMessage {
     final e = this;
-    if (e is DioException && e.response != null) {
-      final body = e.response!.data;
-      String message;
-      try {
-        message = (body as Map)['message'];
-      } catch (_) {
-        message = body;
-      }
-      return '[${e.response!.statusCode}] $message';
+    final (statusCode, message) = switch (this) {
+      RhttpStatusCodeException(:final statusCode, :final body) => (statusCode, _parseErrorMessage(body)),
+      _ => (null, e.toString()),
+    };
+
+    if (statusCode != null && message != null) {
+      return '[$statusCode] $message';
     }
 
     return e.toString();
+  }
+}
+
+String? _parseErrorMessage(Object? body) {
+  if (body is! String) {
+    return null;
+  }
+
+  try {
+    return (jsonDecode(body) as Map)['message'];
+  } catch (_) {
+    return null;
   }
 }