+++ common/lib/src/isolate/child/http_target_discovery_isolate.dart
@@ -1,9 +1,9 @@
 import 'package:common/model/device.dart';
-import 'package:common/src/discovery/http_target_discovery.dart';
 import 'package:common/src/isolate/child/main.dart';
 import 'package:common/src/isolate/dto/isolate_task.dart';
 import 'package:common/src/isolate/dto/isolate_task_result.dart';
 import 'package:common/src/isolate/dto/send_to_isolate_data.dart';
+import 'package:common/src/task/discovery/http_target_discovery.dart';
 import 'package:meta/meta.dart';
 
 class HttpTargetTask {
@@ -21,7 +21,7 @@ class HttpTargetTask {
 @internal
 Future<void> setupHttpTargetDiscoveryIsolate(
   Stream<SendToIsolateData<IsolateTask<HttpTargetTask>>> receiveFromMain,
-  void Function(IsolateTaskResult<Device?>) sendToMain,
+  void Function(IsolateTaskResult<Device>) sendToMain,
   InitialData initialData,
 ) async {
   await setupChildIsolateHelper(
@@ -29,13 +29,23 @@ Future<void> setupHttpTargetDiscoveryIsolate(
     receiveFromMain: receiveFromMain,
     sendToMain: sendToMain,
     initialData: initialData,
-    handler: (task) async {
-      final device = await isolateContainer.read(httpTargetDiscoveryProvider).discover(
+    handler: (ref, task) async {
+      Object? error;
+      final device = await ref.read(httpTargetDiscoveryProvider).discover(
             ip: task.data.ip,
             port: task.data.port,
             https: task.data.https,
+            onError: (url, e) => error = e,
           );
-      sendToMain(IsolateTaskResult(
+
+      if (error != null || device == null) {
+        return sendToMain(IsolateTaskErrorResult(
+          id: task.id,
+          error: error?.toString() ?? 'Unknown error',
+        ));
+      }
+
+      sendToMain(IsolateTaskSuccessResult(
         id: task.id,
         data: device,
       ));