+++ common/lib/src/isolate/child/sync_provider.mapper.dart
@@ -22,6 +22,14 @@ class SyncStateMapper extends ClassMapperBase<SyncState> {
   @override
   final String id = 'SyncState';
 
+  static Function _$init(SyncState v) => (v as dynamic).init as Function;
+  static dynamic _arg$init(f) => f<Future<void> Function()>();
+  static const Field<SyncState, Function> _f$init = Field('init', _$init, arg: _arg$init);
+  static Object _$rootIsolateToken(SyncState v) => v.rootIsolateToken;
+  static const Field<SyncState, Object> _f$rootIsolateToken = Field('rootIsolateToken', _$rootIsolateToken);
+  static Function _$httpClientFactory(SyncState v) => (v as dynamic).httpClientFactory as Function;
+  static dynamic _arg$httpClientFactory(f) => f<CustomHttpClient Function(Duration, StoredSecurityContext)>();
+  static const Field<SyncState, Function> _f$httpClientFactory = Field('httpClientFactory', _$httpClientFactory, arg: _arg$httpClientFactory);
   static StoredSecurityContext _$securityContext(SyncState v) => v.securityContext;
   static const Field<SyncState, StoredSecurityContext> _f$securityContext = Field('securityContext', _$securityContext);
   static DeviceInfoResult _$deviceInfo(SyncState v) => v.deviceInfo;
@@ -43,6 +51,9 @@ class SyncStateMapper extends ClassMapperBase<SyncState> {
 
   @override
   final MappableFields<SyncState> fields = const {
+    #init: _f$init,
+    #rootIsolateToken: _f$rootIsolateToken,
+    #httpClientFactory: _f$httpClientFactory,
     #securityContext: _f$securityContext,
     #deviceInfo: _f$deviceInfo,
     #alias: _f$alias,
@@ -56,6 +67,9 @@ class SyncStateMapper extends ClassMapperBase<SyncState> {
 
   static SyncState _instantiate(DecodingData data) {
     return SyncState(
+        init: data.dec(_f$init),
+        rootIsolateToken: data.dec(_f$rootIsolateToken),
+        httpClientFactory: data.dec(_f$httpClientFactory),
         securityContext: data.dec(_f$securityContext),
         deviceInfo: data.dec(_f$deviceInfo),
         alias: data.dec(_f$alias),
@@ -112,7 +126,10 @@ extension SyncStateValueCopy<$R, $Out> on ObjectCopyWith<$R, SyncState, $Out> {
 abstract class SyncStateCopyWith<$R, $In extends SyncState, $Out> implements ClassCopyWith<$R, $In, $Out> {
   StoredSecurityContextCopyWith<$R, StoredSecurityContext, StoredSecurityContext> get securityContext;
   $R call(
-      {StoredSecurityContext? securityContext,
+      {Future<void> Function()? init,
+      Object? rootIsolateToken,
+      CustomHttpClient Function(Duration, StoredSecurityContext)? httpClientFactory,
+      StoredSecurityContext? securityContext,
       DeviceInfoResult? deviceInfo,
       String? alias,
       int? port,
@@ -134,7 +151,10 @@ class _SyncStateCopyWithImpl<$R, $Out> extends ClassCopyWithBase<$R, SyncState, 
       $value.securityContext.copyWith.$chain((v) => call(securityContext: v));
   @override
   $R call(
-          {StoredSecurityContext? securityContext,
+          {Future<void> Function()? init,
+          Object? rootIsolateToken,
+          CustomHttpClient Function(Duration, StoredSecurityContext)? httpClientFactory,
+          StoredSecurityContext? securityContext,
           DeviceInfoResult? deviceInfo,
           String? alias,
           int? port,
@@ -144,6 +164,9 @@ class _SyncStateCopyWithImpl<$R, $Out> extends ClassCopyWithBase<$R, SyncState, 
           bool? serverRunning,
           bool? download}) =>
       $apply(FieldCopyWithData({
+        if (init != null) #init: init,
+        if (rootIsolateToken != null) #rootIsolateToken: rootIsolateToken,
+        if (httpClientFactory != null) #httpClientFactory: httpClientFactory,
         if (securityContext != null) #securityContext: securityContext,
         if (deviceInfo != null) #deviceInfo: deviceInfo,
         if (alias != null) #alias: alias,
@@ -156,6 +179,9 @@ class _SyncStateCopyWithImpl<$R, $Out> extends ClassCopyWithBase<$R, SyncState, 
       }));
   @override
   SyncState $make(CopyWithData data) => SyncState(
+      init: data.get(#init, or: $value.init),
+      rootIsolateToken: data.get(#rootIsolateToken, or: $value.rootIsolateToken),
+      httpClientFactory: data.get(#httpClientFactory, or: $value.httpClientFactory),
       securityContext: data.get(#securityContext, or: $value.securityContext),
       deviceInfo: data.get(#deviceInfo, or: $value.deviceInfo),
       alias: data.get(#alias, or: $value.alias),