+++ app/assets/CHANGELOG.md
@@ -1,5 +1,16 @@
-## 1.16.0 (unreleased)
+## 1.16.2 (2024-11-06)
 
+- fix(ios): share from other apps to LocalSend doesn't work in iOS 18 (@Tienisto)
+
+## 1.16.1 (2024-11-05)
+
+- feat: show exact error message when using IP address dialog or favorite dialog (@Tienisto)
+- feat(desktop): highlight file when tapping "Show in folder" (@Tienisto)
+- fix(android): properly close app on back gesture (@Tienisto)
+
+## 1.16.0 (2024-11-03)
+
+- feat: improve transfer speed if the sending device is the bottleneck by using Rust as HTTP client and multithreading (@Tienisto)
 - feat: add option to automatically receive files only from favorites (@Davte)
 - feat: only automatically finish when files are either successfully received or skipped (@Tienisto)
 - feat: improve various padding and spacing issues in RTL languages (@ShlomoCode)
@@ -11,9 +22,11 @@
 - feat(macos): starts hidden in menu bar instead of being minimized when autostart is enabled (@ShlomoCode)
 - feat(macos): show error and success state in the app icon (@ShlomoCode, @Tienisto)
 - feat(macos): also have autostart option in sandboxed version (App Store) (@ShlomoCode)
+- feat(macos): LocalSend installed via dmg installer is sandboxed (@Tienisto)
 - feat(android): enable clipboard button (@Seidko)
 - feat(ios): enable clipboard button (@AnessZurba)
 - fix(macos): reopen app from launchpad after minimizing to menu bar should make window visible (@ShlomoCode)
+- fix(macos): persist write access to download location after app restart (@ShlomoCode)
 - i18n: add Malaysian (@Gloridust), Slovak (@dodog)
 
 ## 1.15.4 (2024-08-20)